Преобразование в интерфейс
