+++ src/main/java/school/faang/user_service/service/GoalService.java
@@ -1,20 +1,8 @@
 package school.faang.user_service.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import school.faang.user_service.repository.goal.GoalRepository;
+public interface GoalService {
 
-@Service
-@RequiredArgsConstructor
-public class GoalService {
+    boolean existsById(Long goalId);
 
-    private final GoalRepository goalRepository;
-
-    public boolean existsById(Long goalId) {
-        return goalRepository.existsById(goalId);
-    }
-
-    public long countActiveGoalsPerUser(Long userId) {
-        return goalRepository.countActiveGoalsPerUser(userId);
-    }
+    long countActiveGoalsPerUser(Long userId);
 }
